third_party/blink/web_tests/http/tests/security/subresourceIntegrity/subresource-integrity-style-cors.html

<!DOCTYPE html>
<head>
    <title>SRI on cross origin CORS enabled style</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script>
    setup({single_test: true});

    function styleLoad() {
        assert_true(true, "Style loaded.");
    }
    function styleError() {
        assert_unreached("Style did not load.");
    }
    </script>
    <style>.id1 { background-color: red }</style>
    <link rel="stylesheet"
          onload="styleLoad()"
          onerror="styleError();"
          crossorigin
          integrity="sha256-fy1lC6jbv7rOG6Flb00n5PYLggqe25IcE6QlZf8KtwI="
          href="http://localhost:8000/security/resources/cors-style.php"></link>
</head>
<body>
    <script>
        window.onload = function () {
            var ele = document.getElementById('id1');
            var divBackground = window.getComputedStyle(ele, null).getPropertyValue('background-color');
            assert_equals(divBackground, "rgb(255, 255, 0)");
            done();
        };
    </script>
    <div id="id1" class="id1"></div>
</body>
</html>
